requirements.txt
requests
Pillow
imagehash
praw
python-dotenv
